web-content-change-reminder/requirements.txt
requests
python-dateutil
pytz
python-dotenv
beautifulsoup4